KeeperStorageImpl
